+++ Pysat/Dockerfile.tally
@@ -1,14 +1,9 @@
-FROM with_python as tally
+FROM stevenmburns/pysat_image:2019mar14 as tally_image
+
+COPY tally/ /tally/
 
 RUN \
-    apt-get install zlib1g-dev && \
-    bash -c "source sympy/bin/activate && pip install --upgrade pip && pip install --upgrade wheel && pip install python-sat"
-
-ADD tally/ /tally/
-
-RUN \
-    bash -c "source sympy/bin/activate && cd /tally/ && pip install ." && \
-    rm -r /tally
+    bash -c "source general/bin/activate && cd /tally/ && pip install ."
