+++ Pysat/tally/.gitignore
@@ -1,0 +1,3 @@
+junit.xml
+.eggs/
+.hypothesis/
