+++ Pysat/tally/setup.cfg
@@ -1,0 +1,8 @@
+[aliases]
+test=pytest
+
+[tool:pytest]
+addopts = --junitxml=junit.xml
+
+[tool:pytest:pytest-watch]
+dont-watch-files = ^\.#
