--- Pysat/tally/tally/__init__.py
@@ -1,1 +1,0 @@
-from .tally import *
